+++ android/link-assets-manifest.json
@@ -1,0 +1,53 @@
+{
+  "migIndex": 1,
+  "data": [
+    {
+      "path": "src/assets/fonts/SFProText-Bold.otf",
+      "sha1": "b27d8c428b282fbfd3d40362f4e3ef60afb95ebf"
+    },
+    {
+      "path": "src/assets/fonts/SFProText-BoldItalic.otf",
+      "sha1": "8e96b615ff1f59f8175350483bd897f26e25a278"
+    },
+    {
+      "path": "src/assets/fonts/SFProText-Heavy.otf",
+      "sha1": "330cd932b6dc96b071586409fe36754d5d829955"
+    },
+    {
+      "path": "src/assets/fonts/SFProText-HeavyItalic.otf",
+      "sha1": "f29a62dd89420feb71d609ffb91c470c8492e14a"
+    },
+    {
+      "path": "src/assets/fonts/SFProText-Italic.otf",
+      "sha1": "39e16384b14e45d0c0c002183a1d82bcb74c9897"
+    },
+    {
+      "path": "src/assets/fonts/SFProText-Light.otf",
+      "sha1": "dfc052b7e1aabd14920dbe798a23adcc3a3a521d"
+    },
+    {
+      "path": "src/assets/fonts/SFProText-LightItalic.otf",
+      "sha1": "5c90cf03842aaba33062e8cbea3eb1cee0793748"
+    },
+    {
+      "path": "src/assets/fonts/SFProText-Medium.otf",
+      "sha1": "35b05b13f0c4a63cfc7d23b60c728c5d24cf14fd"
+    },
+    {
+      "path": "src/assets/fonts/SFProText-MediumItalic.otf",
+      "sha1": "2201f9fe8ba96ec444c805d4a8c92849b1e157f2"
+    },
+    {
+      "path": "src/assets/fonts/SFProText-Regular.otf",
+      "sha1": "85b4043974a2a3be5afbcc9ea2becab072d2cd8f"
+    },
+    {
+      "path": "src/assets/fonts/SFProText-Semibold.otf",
+      "sha1": "d8c839390af4265e252fe445b06428c5baeec70d"
+    },
+    {
+      "path": "src/assets/fonts/SFProText-SemiboldItalic.otf",
+      "sha1": "add2bddee4b5699a41ed52c7d30f935d5bb72d14"
+    }
+  ]
+}
