+++ ios/Gooner007ExpenseUAT-Info.plist
@@ -51,5 +51,20 @@
 	</array>
 	<key>UIViewControllerBasedStatusBarAppearance</key>
 	<false/>
+	<key>UIAppFonts</key>
+	<array>
+		<string>SFProText-Bold.otf</string>
+		<string>SFProText-BoldItalic.otf</string>
+		<string>SFProText-Heavy.otf</string>
+		<string>SFProText-HeavyItalic.otf</string>
+		<string>SFProText-Italic.otf</string>
+		<string>SFProText-Light.otf</string>
+		<string>SFProText-LightItalic.otf</string>
+		<string>SFProText-Medium.otf</string>
+		<string>SFProText-MediumItalic.otf</string>
+		<string>SFProText-Regular.otf</string>
+		<string>SFProText-Semibold.otf</string>
+		<string>SFProText-SemiboldItalic.otf</string>
+	</array>
 </dict>
 </plist>
