+++ src/components/typography/H1/index.tsx
@@ -4,18 +4,15 @@ import { Text } from 'react-native'
 import { colors } from '@/constants'
 import { scale } from '@/libs'
 
-const Body: FC<ComponentProps<typeof Text>> = ({
-  children,
-  style,
-  ...rest
-}) => {
+const H1: FC<ComponentProps<typeof Text>> = ({ children, style, ...rest }) => {
   return (
     <Text
       {...rest}
       style={[
         {
           color: colors.black,
-          fontSize: scale(13),
+          fontSize: scale(30),
+          fontFamily: 'SFProText-Bold',
         },
         style,
       ]}>
@@ -24,4 +21,4 @@ const Body: FC<ComponentProps<typeof Text>> = ({
   )
 }
 
-export default Body
+export default H1
